requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
sentencepiece>=0.1.99
protobuf>=4.25.0

# Evaluation
datasets>=2.15.0
evaluate>=0.4.1
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
numpy>=1.24.0
pandas>=2.1.0
tqdm>=4.66.0
pyyaml>=6.0.1
rich>=13.7.0
loguru>=0.7.2

# Statistical analysis
statsmodels>=0.14.0

# Perplexity evaluation
lm-eval>=0.4.0

# Adversarial testing (optional)
# openai>=1.0.0  # For GPT-4 judge

# Development
pytest>=7.4.0
black>=23.12.0
isort>=5.13.0
